refactor: 불필요한 도메인 전부 삭제처리
