+++ src/main/java/balancetalk/game/domain/Game.java
@@ -2,12 +2,9 @@ package balancetalk.game.domain;
 
 import balancetalk.global.common.BaseTimeEntity;
 import balancetalk.module.member.domain.Member;
-import balancetalk.module.vote.domain.Option;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
