+++ src/main/java/balancetalk/game/domain/Option.java
@@ -1,0 +1,5 @@
+package balancetalk.game.domain;
+
+public enum Option {
+    A, B
+}
