--- src/main/java/balancetalk/module/ViewStatus.java
@@ -1,5 +1,0 @@
-package balancetalk.module;
-
-public enum ViewStatus {
-    NORMAL, BLIND
-}
