--- src/main/java/balancetalk/module/bookmark/domain/Bookmark.java
@@ -1,28 +1,0 @@
-package balancetalk.module.bookmark.domain;
-
-import balancetalk.global.common.BaseTimeEntity;
-import balancetalk.module.member.domain.Member;
-import balancetalk.module.post.domain.Post;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Builder
-@Getter
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Bookmark extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "bookmark_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-}
