--- src/main/java/balancetalk/module/comment/domain/Comment.java
@@ -1,54 +1,0 @@
-package balancetalk.module.comment.domain;
-
-import balancetalk.global.common.BaseTimeEntity;
-import balancetalk.module.ViewStatus;
-import balancetalk.module.member.domain.Member;
-import balancetalk.module.post.domain.Post;
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-import lombok.*;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Builder
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class Comment extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "comment_id")
-    private Long id;
-
-    @NotBlank
-    @Size(max = 100)
-    private String content;
-
-    @Enumerated(value = EnumType.STRING)
-    @NotNull
-    private ViewStatus viewStatus;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id") // TODO : post_id -> 톡픽_id
-    private Post post;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "parent_id")
-    private Comment parent;
-
-    @OneToMany(mappedBy = "comment", cascade = CascadeType.ALL)
-    private List<CommentLike> likes = new ArrayList<>();
-
-    public void updateContent(String content) {
-        this.content = content;
-    }
-}
