--- src/main/java/balancetalk/module/comment/domain/CommentLike.java
@@ -1,34 +1,0 @@
-package balancetalk.module.comment.domain;
-
-import balancetalk.module.member.domain.Member;
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotBlank;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class CommentLike {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "comment_like_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "comment_id")
-    private Comment comment;
-
-    @NotBlank
-    private boolean isActive;
-}
