--- src/main/java/balancetalk/module/comment/domain/CommentLikeRepository.java
@@ -1,12 +1,0 @@
-package balancetalk.module.comment.domain;
-
-import balancetalk.module.member.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CommentLikeRepository extends JpaRepository<CommentLike, Long> {
-
-
-    boolean existsByMemberAndComment(Member member, Comment comment);
-
-    void deleteByMemberAndComment(Member member, Comment comment);
-}
