+++ src/main/java/balancetalk/module/member/domain/MemberRepository.java
@@ -12,6 +12,6 @@ public interface MemberRepository extends JpaRepository<Member, Long> {
     boolean existsByEmail(String email);
     void deleteByEmail(String email);
 
-    @Query("select m.id from Member m JOIN m.votes v WHERE v.balanceOption.id = :balanceOptionId")
-    List<Long> findMemberIdsBySelectedOptionId(Long balanceOptionId);
+//    @Query("select m.id from Member m JOIN m.votes v WHERE v.balanceOption.id = :balanceOptionId")
+//    List<Long> findMemberIdsBySelectedOptionId(Long balanceOptionId);
 }
