--- src/main/java/balancetalk/module/post/domain/BalanceOptionRepository.java
@@ -1,6 +1,0 @@
-package balancetalk.module.post.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BalanceOptionRepository extends JpaRepository<BalanceOption, Long> {
-}
