--- src/main/java/balancetalk/module/post/domain/PostCategory.java
@@ -1,5 +1,0 @@
-package balancetalk.module.post.domain;
-
-public enum PostCategory {
-    CASUAL, DISCUSSION
-}
