--- src/main/java/balancetalk/module/post/domain/PostLikeRepository.java
@@ -1,11 +1,0 @@
-package balancetalk.module.post.domain;
-
-import balancetalk.module.member.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostLikeRepository extends JpaRepository<PostLike, Long> {
-
-    boolean existsByMemberAndPost(Member member, Post post);
-
-    void deleteByMemberAndPost(Member member, Post post);
-}
