--- src/main/java/balancetalk/module/post/domain/PostTag.java
@@ -1,33 +1,0 @@
-package balancetalk.module.post.domain;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Builder
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class PostTag {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "post_tag_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-    @JoinColumn(name = "tag_id")
-    private Tag tag;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
-
-    @Builder
-    public PostTag(Tag tag) {
-        this.tag = tag;
-    }
-    public void addPost(Post post) {
-        this.post = post;
-    }
-}
