--- src/main/java/balancetalk/module/post/domain/Tag.java
@@ -1,30 +1,0 @@
-package balancetalk.module.post.domain;
-
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
-import lombok.*;
-
-import java.util.ArrayList;
-import java.util.List;
-@Entity
-@Builder
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class Tag {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "tag_id")
-    private Long id;
-
-    @NotNull
-    private String name;
-
-    @OneToMany(mappedBy = "tag")
-    private List<PostTag> postTags = new ArrayList<>();
-
-    public Tag(String name){
-        this.name = name;
-    }
-}
