--- src/main/java/balancetalk/module/vote/domain/Option.java
@@ -1,5 +1,0 @@
-package balancetalk.module.vote.domain;
-
-public enum Option {
-    A, B
-}
