--- src/main/java/balancetalk/module/vote/domain/Vote.java
@@ -1,33 +1,0 @@
-package balancetalk.module.vote.domain;
-
-import balancetalk.module.member.domain.Member;
-import balancetalk.module.post.domain.BalanceOption;
-import balancetalk.global.common.BaseTimeEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Builder
-@Getter
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Vote extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "vote_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "balance_option_id")
-    private BalanceOption balanceOption;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    public Vote changeBalanceOption(BalanceOption newSelectedOption) {
-        this.balanceOption = newSelectedOption;
-        return this;
-    }
-}
